Adding time travel for navigating to previous moves
A list of previous moves is displayed on the right and added to whenever a move is made. If a player wishes to go back to a previous move, they need only click the appropriate list button and the board reverts to the state of the board after that move.